+++ src/br/com/resource/catalogoconhecimento/business/NegocioBusiness.java
@@ -7,6 +7,7 @@ import br.com.resource.catalogoconhecimento.dao.NegocioDAO;
 import br.com.resource.catalogoconhecimento.exceptions.AtributoNuloException;
 import br.com.resource.catalogoconhecimento.exceptions.ConsultaNulaException;
 import br.com.resource.catalogoconhecimento.exceptions.NomeRepetidoException;
+import br.com.resource.catalogoconhecimento.exceptions.RegistroVinculadoException;
 import br.com.resource.catalogoconhecimento.exceptions.TamanhoCampoException;
 
 public class NegocioBusiness {
@@ -74,10 +75,17 @@ public class NegocioBusiness {
 		}
 	}
 	
-	public void remover(int id) throws ClassNotFoundException, SQLException {
+	public void remover(int id) throws ClassNotFoundException, SQLException, RegistroVinculadoException {
 		NegocioDAO negocioDao = new NegocioDAO();
+		if(negocioDao.verificarPorProjeto(id)){
+			negocioDao.remover(id);	
+		} else {
+			throw new RegistroVinculadoException("Registro não pode ser removido pois possui vínculos");
+			
+			
+		}
 		
-		negocioDao.remover(id);
+		
 	}
 	
 	public void reativar(NegocioBean negocioBean) throws ClassNotFoundException, SQLException{
